feat: 경기일정 선택 (#36)
