+++ src/apis/main/index.ts
@@ -1,6 +1,8 @@
 import { defaultApi } from "../core/index";
 import { Schedule } from "../../components/home/Card";
 
+const getAuthToken = () => localStorage.getItem("token") || "";
+
 export const home = {
   place: async (stadium: string, category: string) => {
     try {
@@ -32,7 +34,7 @@ export const home = {
 };
 
 export const fetchSchedules = async (team: string): Promise<Schedule[]> => {
-  const token = localStorage.getItem("token") || "";
+  const token = getAuthToken();
   try {
     const response = await defaultApi.get<{ schedules: Schedule[] }>(
       `/api/main/schedule/?team=${encodeURIComponent(team)}&page=0&size=50`,
@@ -50,8 +52,9 @@ export const fetchSchedules = async (team: string): Promise<Schedule[]> => {
 };
 
 export const scrapSchedule = async (gameId: number) => {
-  const token = localStorage.getItem("token") || "";
+  const token = getAuthToken();
   try {
+    console.log("Scrapping schedule for gameId:", gameId);
     const response = await defaultApi.patch<{}>(
       `/api/scraps/schedule/scrap?gameId=${gameId}`,
       {},
