+++ src/components/home/Card.tsx
@@ -10,19 +10,7 @@ import { toast } from "react-toastify";
 import { fetchSchedules } from "../../apis/main";
 import { scrapSchedule } from "../../apis/main";
 import useTeamStore from "../../store/TeamStore";
-
-export const teamLogos: Record<string, string> = {
-  LG: "https://yaguhang.kro.kr:8443/teamLogos/LGTwins.png",
-  KT: "https://yaguhang.kro.kr:8443/teamLogos/KtWizs.png",
-  SSG: "https://yaguhang.kro.kr:8443/teamLogos/SSGLanders.png",
-  NC: "https://yaguhang.kro.kr:8443/teamLogos/NCDinos.png",
-  두산: "https://yaguhang.kro.kr:8443/teamLogos/Doosan.png",
-  KIA: "https://yaguhang.kro.kr:8443/teamLogos/KIA.png",
-  롯데: "https://yaguhang.kro.kr:8443/teamLogos/Lotte.png",
-  삼성: "https://yaguhang.kro.kr:8443/teamLogos/Samsung.png",
-  한화: "https://yaguhang.kro.kr:8443/teamLogos/Hanwha.png",
-  키움: "https://yaguhang.kro.kr:8443/teamLogos/Kiwoom.png",
-};
+import { teamLogos } from "../../types/teamLogos";
 
 export interface Schedule {
   id: number;
@@ -41,6 +29,158 @@ export interface Schedule {
 interface StyledCardProps {
   $isScraped: boolean;
 }
+
+const Card: React.FC = () => {
+  const [schedules, setSchedules] = useState<Schedule[]>([]);
+  const [currentPage, setCurrentPage] = useState(0);
+  const { selectedTeam, setSelectedGame, selectedGame } = useTeamStore();
+
+  useEffect(() => {
+    const loadSchedules = async () => {
+      const schedules = await fetchSchedules(selectedTeam);
+      setSchedules(schedules);
+      setCurrentPage(0); // 팀이 변경될 때 페이지를 초기화
+    };
+    loadSchedules();
+  }, [selectedTeam]); // selectedTeam이 변경될 때마다 경기일정 필터링
+
+  const schedulesPerPage = 5;
+  const indexOfLastSchedule = (currentPage + 1) * schedulesPerPage;
+  const indexOfFirstSchedule = indexOfLastSchedule - schedulesPerPage;
+  const currentSchedules = schedules.slice(
+    indexOfFirstSchedule,
+    indexOfLastSchedule
+  );
+
+  const nextPage = () => {
+    if (indexOfLastSchedule < schedules.length) {
+      setCurrentPage(currentPage + 1);
+    }
+  };
+
+  const prevPage = () => {
+    if (currentPage > 0) {
+      setCurrentPage(currentPage - 1);
+    }
+  };
+
+  const handleScrapSchedule = async (gameId: number) => {
+    try {
+      const isScraped = await scrapSchedule(gameId);
+      setSchedules((prevSchedules) =>
+        prevSchedules.map((schedule) =>
+          schedule.id === gameId
+            ? { ...schedule, isScraped: !schedule.isScraped }
+            : schedule
+        )
+      );
+      toast.success(
+        isScraped ? "스크랩에 추가되었습니다." : "스크랩에서 제거되었습니다."
+      );
+    } catch (error) {
+      toast.error("스크랩 중 오류가 발생했습니다.");
+    }
+  };
+
+  //카드 클릭시 경기 날짜, 경기장 저장
+  const handleCardClick = (schedule: Schedule) => {
+    setSelectedGame({ date: schedule.date, stadium: schedule.stadium });
+    console.log("선택된 게임:", {
+      date: schedule.date,
+      stadium: schedule.stadium,
+    });
+  };
+
+  return (
+    <>
+      <Category filterSchedules={fetchSchedules} teamLogos={teamLogos} />{" "}
+      <S.Wrapper gap="100px">
+        <S.TitleWrapper>
+          <S.Title style={{ color: "#ffffff" }}>오늘의 경기 일정</S.Title>
+          <S.H4>
+            열정과 감동을 선사하는 스포츠! 그 현장을 직접 경험해보세요!
+          </S.H4>
+        </S.TitleWrapper>
+      </S.Wrapper>
+      <CardContainer>
+        <PrevButton onClick={prevPage} disabled={currentPage === 0}>
+          <img src={left} alt="이전" />
+        </PrevButton>
+        {currentSchedules.map((schedule) => (
+          <StyledCard
+            key={schedule.id}
+            $isScraped={schedule.isScraped}
+            onClick={() => handleCardClick(schedule)}
+            style={{
+              border:
+                selectedGame?.date === schedule.date &&
+                selectedGame?.stadium === schedule.stadium
+                  ? "2px solid #FF0000"
+                  : "1px solid #ffffff",
+            }}
+          >
+            <BeforeElement
+              $isScraped={schedule.isScraped}
+              onClick={() => handleScrapSchedule(schedule.id)}
+            />
+            <div style={{ marginTop: "2rem" }}>
+              <div>
+                {schedule.stadium} | {schedule.time}
+              </div>
+              <div
+                style={{
+                  display: "flex",
+                  justifyContent: "center",
+                  alignItems: "center",
+                  marginTop: "2.5vh",
+                }}
+              >
+                <img
+                  src={schedule.homeTeamLogo}
+                  alt={`${schedule.home}`}
+                  style={{ width: "3rem", height: "2.5rem" }}
+                />
+                <span style={{ margin: "0 1rem" }}>VS</span>
+                <img
+                  src={schedule.awayTeamLogo}
+                  alt={`${schedule.away}`}
+                  style={{ width: "3rem", height: "2.5rem" }}
+                />
+              </div>
+              <div
+                style={{
+                  display: "flex",
+                  justifyContent: "space-around",
+                  marginTop: "10px",
+                  fontSize: "0.8rem",
+                }}
+              >
+                <span style={{ whiteSpace: "pre-line" }}>{schedule.home}</span>
+                <span style={{ whiteSpace: "pre-line" }}>{schedule.away}</span>
+              </div>
+              <Divider />
+              <DateAndWeatherContainer>
+                <DateContainer>{schedule.date}</DateContainer>
+                <WeatherIcon
+                  src={schedule.weatherUrl}
+                  alt={`${schedule.weather}`}
+                />
+              </DateAndWeatherContainer>
+            </div>
+          </StyledCard>
+        ))}
+        <NextButton
+          onClick={nextPage}
+          disabled={indexOfLastSchedule >= schedules.length}
+        >
+          <img src={right} alt="다음" />
+        </NextButton>
+      </CardContainer>
+    </>
+  );
+};
+
+export default Card;
 
 const CardContainer = styled.div`
   display: flex;
@@ -123,7 +263,6 @@ const PaginationButton = styled.button`
     height: 2rem;
   }
 `;
-
 const PrevButton = styled(PaginationButton)`
   left: 12vw;
 `;
@@ -131,135 +270,3 @@ const PrevButton = styled(PaginationButton)`
 const NextButton = styled(PaginationButton)`
   right: 12vw;
 `;
-
-const Card: React.FC = () => {
-  const [schedules, setSchedules] = useState<Schedule[]>([]);
-  const [currentPage, setCurrentPage] = useState(0);
-  const { selectedTeam } = useTeamStore();
-
-  useEffect(() => {
-    const loadSchedules = async () => {
-      const schedules = await fetchSchedules(selectedTeam);
-      setSchedules(schedules);
-      setCurrentPage(0); // 팀이 변경될 때 페이지를 초기화
-    };
-    loadSchedules();
-  }, [selectedTeam]); // selectedTeam이 변경될 때마다 경기일정 필터링
-
-  const schedulesPerPage = 5;
-  const indexOfLastSchedule = (currentPage + 1) * schedulesPerPage;
-  const indexOfFirstSchedule = indexOfLastSchedule - schedulesPerPage;
-  const currentSchedules = schedules.slice(
-    indexOfFirstSchedule,
-    indexOfLastSchedule
-  );
-
-  const nextPage = () => {
-    if (indexOfLastSchedule < schedules.length) {
-      setCurrentPage(currentPage + 1);
-    }
-  };
-
-  const prevPage = () => {
-    if (currentPage > 0) {
-      setCurrentPage(currentPage - 1);
-    }
-  };
-
-  const handleScrapSchedule = async (gameId: number) => {
-    try {
-      const isScraped = await scrapSchedule(gameId);
-      setSchedules((prevSchedules) =>
-        prevSchedules.map((schedule) =>
-          schedule.id === gameId
-            ? { ...schedule, isScraped: !schedule.isScraped }
-            : schedule
-        )
-      );
-      toast.success(
-        isScraped ? "스크랩에 추가되었습니다." : "스크랩에서 제거되었습니다."
-      );
-    } catch (error) {
-      toast.error("스크랩 중 오류가 발생했습니다.");
-    }
-  };
-
-  return (
-    <>
-      <Category filterSchedules={fetchSchedules} teamLogos={teamLogos} />{" "}
-      <S.Wrapper gap="100px">
-        <S.TitleWrapper>
-          <S.Title style={{ color: "#ffffff" }}>오늘의 경기 일정</S.Title>
-          <S.H4>
-            열정과 감동을 선사하는 스포츠! 그 현장을 직접 경험해보세요!
-          </S.H4>
-        </S.TitleWrapper>
-      </S.Wrapper>
-      <CardContainer>
-        <PrevButton onClick={prevPage} disabled={currentPage === 0}>
-          <img src={left} alt="이전" />
-        </PrevButton>
-        {currentSchedules.map((schedule) => (
-          <StyledCard key={schedule.id} $isScraped={schedule.isScraped}>
-            <BeforeElement
-              $isScraped={schedule.isScraped}
-              onClick={() => handleScrapSchedule(schedule.id)}
-            />
-            <div style={{ marginTop: "2rem" }}>
-              <div>
-                {schedule.stadium} | {schedule.time}
-              </div>
-              <div
-                style={{
-                  display: "flex",
-                  justifyContent: "center",
-                  alignItems: "center",
-                  marginTop: "2.5vh",
-                }}
-              >
-                <img
-                  src={schedule.homeTeamLogo}
-                  alt={`${schedule.home}`}
-                  style={{ width: "3rem", height: "2.5rem" }}
-                />
-                <span style={{ margin: "0 1rem" }}>VS</span>
-                <img
-                  src={schedule.awayTeamLogo}
-                  alt={`${schedule.away}`}
-                  style={{ width: "3rem", height: "2.5rem" }}
-                />
-              </div>
-              <div
-                style={{
-                  display: "flex",
-                  justifyContent: "space-around",
-                  marginTop: "10px",
-                  fontSize: "0.8rem",
-                }}
-              >
-                <span style={{ whiteSpace: "pre-line" }}>{schedule.home}</span>
-                <span style={{ whiteSpace: "pre-line" }}>{schedule.away}</span>
-              </div>
-              <Divider />
-              <DateAndWeatherContainer>
-                <DateContainer>{schedule.date}</DateContainer>
-                <WeatherIcon
-                  src={schedule.weatherUrl}
-                  alt={`${schedule.weather}`}
-                />
-              </DateAndWeatherContainer>
-            </div>
-          </StyledCard>
-        ))}
-        <NextButton
-          onClick={nextPage}
-          disabled={indexOfLastSchedule >= schedules.length}
-        >
-          <img src={right} alt="다음" />
-        </NextButton>
-      </CardContainer>
-    </>
-  );
-};
-
-export default Card;
