+++ src/types/teamLogos.ts
@@ -1,0 +1,12 @@
+export const teamLogos: Record<string, string> = {
+  LG: "https://yaguhang.kro.kr:8443/teamLogos/LGTwins.png",
+  KT: "https://yaguhang.kro.kr:8443/teamLogos/KtWizs.png",
+  SSG: "https://yaguhang.kro.kr:8443/teamLogos/SSGLanders.png",
+  NC: "https://yaguhang.kro.kr:8443/teamLogos/NCDinos.png",
+  두산: "https://yaguhang.kro.kr:8443/teamLogos/Doosan.png",
+  KIA: "https://yaguhang.kro.kr:8443/teamLogos/KIA.png",
+  롯데: "https://yaguhang.kro.kr:8443/teamLogos/Lotte.png",
+  삼성: "https://yaguhang.kro.kr:8443/teamLogos/Samsung.png",
+  한화: "https://yaguhang.kro.kr:8443/teamLogos/Hanwha.png",
+  키움: "https://yaguhang.kro.kr:8443/teamLogos/Kiwoom.png",
+};
